refactor tests and add segmentation local upload test
