+++ tests/test_annotation.py
@@ -34,7 +34,6 @@ from .helpers import (
     assert_multicategory_annotation_matches_dict,
     assert_partial_equality,
     assert_polygon_annotation_matches_dict,
-    assert_segmentation_annotation_matches_dict,
     reference_id_from_url,
 )
 
@@ -240,70 +239,6 @@ def test_default_multicategory_gt_upload(dataset):
     assert_multicategory_annotation_matches_dict(
         response_annotation, TEST_DEFAULT_MULTICATEGORY_ANNOTATIONS[0]
     )
-
-
-def test_single_semseg_gt_upload(dataset):
-    annotation = SegmentationAnnotation.from_json(
-        TEST_SEGMENTATION_ANNOTATIONS[0]
-    )
-    response = dataset.annotate(annotations=[annotation])
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 1
-    assert response["annotations_ignored"] == 0
-
-    response_annotation = dataset.refloc(annotation.reference_id)[
-        "annotations"
-    ]["segmentation"][0]
-    assert_segmentation_annotation_matches_dict(
-        response_annotation, TEST_SEGMENTATION_ANNOTATIONS[0]
-    )
-
-
-def test_batch_semseg_gt_upload(dataset):
-    annotations = [
-        SegmentationAnnotation.from_json(ann)
-        for ann in TEST_SEGMENTATION_ANNOTATIONS
-    ]
-    response = dataset.annotate(annotations=annotations)
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 5
-    assert response["annotations_ignored"] == 0
-
-
-def test_batch_semseg_gt_upload_ignore(dataset):
-    # First upload annotations
-    annotations = [
-        SegmentationAnnotation.from_json(ann)
-        for ann in TEST_SEGMENTATION_ANNOTATIONS
-    ]
-    response = dataset.annotate(annotations=annotations)
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 5
-    assert response["annotations_ignored"] == 0
-
-    # When we re-upload, expect them to be ignored
-    response = dataset.annotate(annotations=annotations)
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 0
-    assert response["annotations_ignored"] == 5
-
-
-def test_batch_semseg_gt_upload_update(dataset):
-    # First upload annotations
-    annotations = [
-        SegmentationAnnotation.from_json(ann)
-        for ann in TEST_SEGMENTATION_ANNOTATIONS
-    ]
-    response = dataset.annotate(annotations=annotations)
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 5
-    assert response["annotations_ignored"] == 0
-
-    # When we re-upload, expect uploads to be processed
-    response = dataset.annotate(annotations=annotations, update=True)
-    assert response["dataset_id"] == dataset.id
-    assert response["annotations_processed"] == 5
-    assert response["annotations_ignored"] == 0
 
 
 def test_mixed_annotation_upload(dataset):
